Fixed compiler warnings in main/

+++ main/main.c
@@ -1621,7 +1621,7 @@ void php_request_shutdown(void *dummy)
 		zend_bool send_buffer = SG(request_info).headers_only ? 0 : 1;
 
 		if (CG(unclean_shutdown) && PG(last_error_type) == E_ERROR &&
-			PG(memory_limit) < zend_memory_usage(1 TSRMLS_CC)
+			(size_t)PG(memory_limit) < zend_memory_usage(1 TSRMLS_CC)
 		) {
 			send_buffer = 0;
 		}
